Improved error handling and fixed progress_bar issues.
1- self.make_soup() would fail because of unreturned recursive
error handling. So in this commit this is fixed by returning all
recurse calls in handle_network_errors() method.

2- The functionality of asking the user if a file exists didn't work
because in rich you can't use the input() function inside a
rich.Progress. So now each task that needs a progress_bar will
initialize it's own rather than accepting one global Progress and adding
tasks to that.