Fixed bug occured when exiting when downloading
In utils.handle_keyboard_interrupt_for_files the code would
raise a typer.Exit error when it finds one but the problem is
the function that is decorated runs in utils.handle_network_errors
which has its own exception handling technique. So fixed the bug
by excepting the typer.Exit error and raising it again for typer
to catch it.